+++ partiql-eval/src/main/kotlin/org/partiql/eval/internal/operator/rex/ExprPathSymbol.kt
@@ -13,9 +13,6 @@ internal class ExprPathSymbol(
 
     override fun eval(env: Environment): Datum {
         val struct = root.eval(env).checkStruct()
-        if (struct.isMissing) {
-            return Datum.missing()
-        }
         if (struct.isNull) {
             return Datum.nullValue()
         }